Major cleanup - added C++11 support - dropped current jasstrans code - all scoped enums are actual scoped enums now - dopped editor::Plugin - added separate class editor::WarcraftIIIShared - moved stuff from editor::MpqPriorityList to editor::WarcraftIIIShared - dropped editor::MpqArchive class - cleaned up class Exception - fixed size types in utility stream functions - replaced boost::array by std::array - improved basic vertex classes

+++ src/map/environment.cpp
@@ -98,7 +98,7 @@ std::streamsize Environment::read(InputStream &istream) throw (class Exception)
 std::streamsize Environment::write(OutputStream &ostream) const throw (class Exception)
 {
 	std::streamsize size = FileFormat::write(ostream);
-	wc3lib::write<byte>(ostream, mainTileset(), size);
+	wc3lib::write<byte>(ostream, static_cast<byte>(mainTileset()), size);
 	wc3lib::write<uint32>(ostream, customized(), size);
 
 	if (groundTilesetsIds().size() > maxTilesets)
